backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
blis==0.7.11
Brlapi==0.8.3
catalogue==2.0.10
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloudpathlib==0.16.0
cmake==3.29.3
colorama==0.4.4
command-not-found==0.3
confection==0.1.4
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
easyocr==1.7.1
email_validator==2.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.14.1
filelock==3.14.0
fonttools==4.52.4
fsspec==2024.5.0
future==0.18.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.3
imageio==2.34.1
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language-selector==0.1
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lit==18.1.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.9.0
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
murmurhash==1.0.10
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.9.0.80
orjson==3.10.3
packaging==24.0
paramiko==2.9.3
pexpect==4.8.0
pillow==10.3.0
preshed==3.0.9
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pyclipper==1.3.0.post5
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.7.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-triton-rocm==2.3.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==13.7.1
scikit-image==0.23.2
scipy==1.13.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.11
starlette==0.37.2
sympy==1.12
systemd-python==234
thinc==8.2.3
tifffile==2024.5.22
torch==2.3.0+cu121
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu121
tqdm==4.66.4
triton==2.3.0
typer==0.9.4
typing_extensions==4.12.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.6
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
xdg==5
xkit==0.0.0
zipp==1.0.0
